+++ co-kkiri/src/components/commons/DropDowns/DeadlineDropdown.tsx
@@ -1,18 +1,18 @@
 import styled from "styled-components";
-import DropdownButton from "./commons/DropdownButton";
 import { useState } from "react";
 import useOpenToggle from "@/hooks/useOpenToggle";
 import DESIGN_TOKEN from "@/styles/tokens";
 import { DatePickerWithRange } from "../DatePickerWithRange";
+import SquareDropButton from "./commons/SquareDropButton";
 
-export default function DeadlineDropdownButton() {
-  const [selectOption, setSelectOption] = useState();
+export default function DeadlineDropdown() {
+  const [selectOption, setSelectOption] = useState("마감 기간");
   const [isSelected, setIsSelected] = useState(false);
   const { isOpen, openToggle: toggleDropdown, ref } = useOpenToggle();
 
   return (
     <Container ref={ref}>
-      <DropdownButton $selectType="date" onClick={toggleDropdown} selectOption="마감 기간" $isSelected={isSelected} />
+      <SquareDropButton iconType="date" onClick={toggleDropdown} selectOption={selectOption} $isSelected={isSelected} />
       {isOpen && <DatePickerWithRange />}
     </Container>
   );
