+++ co-kkiri/src/components/commons/DropDowns/FilterDropdown.tsx
@@ -2,45 +2,41 @@ import { useState } from "react";
 import styled from "styled-components";
 import FilterButton from "../FilterButton";
 import useOpenToggle from "@/hooks/useOpenToggle";
-import DropdownMenu from "./commons/DropdownMenu";
+import DropMenu from "./commons/DropMenu";
+import { DROPDOWN_INFO } from "@/constants/dropDown";
 
-interface FilterDropdownButtonProps {
-  filterType: "position" | "progressWay";
-  onSelectFilter: (filterType: string, selectedValue: string) => void;
+interface FilterDropdownProps {
+  menuInfoType: "position" | "progressWay";
+  onSelectFilter: (menuInfoType: string, selectedValue: string) => void;
 }
 
-const filterDefaultValue = {
-  position: "포지션",
-  progressWay: "진행 방식",
-};
-
 /**
- * FilterDropdownButton 컴포넌트
- * filterType에 따라 dropdown 옵션이 정해집니다.
- * @property {"position"|"progressWay"} filterType
+ * FilterDropdown 컴포넌트
+ * @param menuInfoType: 드랍메뉴 내용
+ * @property {"position"|"progressWay"} menuInfoType
  * */
-export default function FilterDropdownButton({ filterType, onSelectFilter }: FilterDropdownButtonProps) {
-  const defaultValue = filterDefaultValue[filterType];
+export default function FilterDropdown({ menuInfoType, onSelectFilter }: FilterDropdownProps) {
+  const { filter } = DROPDOWN_INFO;
 
-  const [selectOption, setSelectOption] = useState(defaultValue);
+  const [selectOption, setSelectOption] = useState(filter[menuInfoType].defaultValue);
   const [isSelected, setIsSelected] = useState(false);
   const { isOpen, openToggle: toggleDropdown, ref } = useOpenToggle();
 
   const handleSelectOption = (option: string) => {
     setSelectOption(option);
     setIsSelected(true);
-    onSelectFilter(filterType, option);
+    onSelectFilter(menuInfoType, option);
     toggleDropdown();
   };
 
   return (
     <Container ref={ref}>
       <FilterButton onClick={toggleDropdown} selectOption={selectOption} isSelected={isSelected} />
-      <DropdownMenu
+      <DropMenu
         $borderType="round"
         isOpen={isOpen}
         handleSelectOption={handleSelectOption}
-        $selectType={filterType}
+        options={filter[menuInfoType].options}
       />
     </Container>
   );
